requirements.txt
Django==3.1.5
django-ckeditor==6.0.0
Pillow==8.1.0
django-model-utils==4.1.1
django-environ==0.4.5
factory-boy==3.2.0
Faker==5.8.0
